sentinel/requirements.txt
sentinelhub
python-dotenv
numpy
matplotlib
